Client-side permissions UI
Most the change are relating to passing down `tagEditable` state. This
ensures the user sees disabled input elements indicating the tag is
readonly. To make sure the user is fully aware of the readonly status
there is a great big orange bar stating they don't have permission to
edit this tag.

There is also a fail-safe in that even if the user does manage to edit
a field in the client, the save bar will not appear. Of course the
server side permissions checking is still in action so if, for example,
the clients permissions cache is out of date with the servers then the
server will reject the save and the client will see a red error bar
indicatng they couldnt save.

Note: Generally we pass `tagEditable` to our custom react compontents,
rather than disabled. This is because the component can still perform
some actions even if the tag is not editable (such as render readonly
content). However, there are a few exceptions namely components in the
public/components/utils directory. These are supposed to be generic and
reusable and should hence have no concept of what a tag is.