Uses play json extensions for Tag json format as it's now bigger that 22 fields
Post merge the tag case class has grown to over 22 fields, this is supported since scala 2.11 but the underlying functionN methods that are used to power the default play json format stuff means that the json stuff broke. This adds https://github.com/cvogt/play-json-extensions to get around the limit, which is nice. It also massively simplified writing the formats for case classes and supports default values and singe value case classes. result all round.

To make it work i had to make the internal case class field names match the json format, but rem=naming references to external references is not such a bad change.