remove code in denormalised tag normalisation process to update sponsorship state
The code was updating the sponsorship state and the tag's active sponsorships but was not performing the section update or content expires. The code has been removed to let the sponsorship expiry and launch jobs handle paid content sponsorship state changes, thus there is only one place this happens and the confusing incomplete launch by side effect is removed.