Do consistent reads out of the sponsorship repository
When creating a tag with a sponsorship, we'll do a read out of the sponsorship repository very soon after writing. Usually (99% of the time) that won't cause any problem whatsoever. But DynamoDB is *eventually* consistent by default, which means that *sometimes* the sponsorship won't be returned on a Get. This means that the active sponsorship doesn't get added to the update sent to CAPI, which can cause very confusing behaviour downstream.

I think that the need to read back out of the sponsorship repository is a weakness in the mapping between Thrift and Scala models, and can probably be refactored out, but running the GetItem in consistent reads mode makes a good start, and is probably good practice.