Username is passed around as string not as `User` object
Batch tags are currently always logged as "default user" because this function never receives an implicit User object.

It feels unpleasant to add another implicit String parameter, since this means the function is effectively saying to its caller "Hey, do you have a string handy that I can pretend is a username?". It would be nicer if this was passed as a User object.

This issue also highlights how easy it is for bugs to creep in with implicit parameters that have default values.