Pre-Scala 2.13: Upgrade deps to versions that are also available on Scala 2.13
As preparation for Scala 2.13, this updates all dependencies so that they
are at versions that are available for Scala 2.13 as well as Scala 2.12.

As ScalaTest is updated to ScalaTest v3.2, we had to do some code updates
to cope with class/package changes. We did these manually, though later
we remembered that maybe we could have used Scalafix, as in this example here:

* https://github.com/guardian/content-api/pull/2893

The steps would have been:

- add sbt-scalafix plugin to user-level sbt plugins file
  - addSbtPlugin("ch.epfl.scala" % "sbt-scalafix" % "0.12.1")
- checked out main branch, still scala 2.12 (and so already compiled successfully)
- launch sbt
- run `scalafixEnable`
- run `scalafixAll dependency:RewriteDeprecatedNames@org.scalatest:autofix:3.1.0.1`
- increment scalatest dependency